Move db provider methods to a separate file
