+++ user_svc/auth/user_manager.py
@@ -1,11 +1,10 @@
 import uuid
 from typing import Optional
 
-from fastapi import Depends, Request
+from fastapi import Request
 from fastapi_users import BaseUserManager, UUIDIDMixin
 
 from db.models import User
-from db.session import get_user_db
 
 from utils.logger import log
 
@@ -24,7 +23,3 @@ class UserManager(UUIDIDMixin, BaseUserManager[User, uuid.UUID]):
         self, user: User, token: str, request: Optional[Request] = None
     ):
         log.info(f"Verification requested for user {user.id}. Verification token: {token}")
-
-
-async def get_user_manager(user_db=Depends(get_user_db)):
-    yield UserManager(user_db)
